requirements.txt
gspread==0.6.2
requests==2.9.1
google_api_python_client==1.6.4
Flask==0.12.2
GitPython==2.1.8
httplib2==0.10.3
PyYAML==3.12
